requirements.txt
aiobotocore==2.13.1
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
basemap==1.4.1
basemap-data==1.3.2
blinker==1.8.2
botocore==1.34.131
Cartopy==0.23.0
certifi==2024.6.2
cftime==1.6.4
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.2.1
cycler==0.12.1
fiona==1.9.6
Flask==3.0.3
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
geojson==3.1.0
geopandas==0.14.4
geos==0.2.3
h5netcdf==1.3.0
h5py==3.11.0
idna==3.7
imageio==2.34.1
importlib_metadata==7.2.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.5
lazy_loader==0.4
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.8.4
multidict==6.0.5
netCDF4==1.7.1
networkx==3.2.1
numpy==1.26.4
packaging==23.2
pandas==2.2.2
pillow==10.3.0
PyMySQL==1.1.1
pyparsing==3.1.2
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.6.0
redis==5.0.6
requests==2.32.3
s3fs==2024.6.0
s3transfer==0.10.1
scikit-image==0.24.0
scipy==1.13.1
shapely==2.0.4
six==1.16.0
tifffile==2024.6.18
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
Werkzeug==3.0.3
wrapt==1.16.0
xarray==2024.6.0
yarl==1.9.4
zipp==3.19.2
